+++ http4k-multipart/src/main/kotlin/org/http4k/core/deprecated.kt
@@ -1,0 +1,6 @@
+package org.http4k.core
+
+import org.http4k.lens.MultipartFormFile
+
+@Deprecated("use ", ReplaceWith("org.http4k.lens.MultipartFormFile"))
+typealias FormFile = MultipartFormFile