+++ src/docs/cookbook/multipart_forms/example_lens.kt
@@ -3,7 +3,6 @@ package cookbook.multipart_forms
 import org.http4k.client.ApacheClient
 import org.http4k.core.Body
 import org.http4k.core.ContentType
-import org.http4k.core.FormFile
 import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.Response
@@ -16,7 +15,6 @@ import org.http4k.lens.MultipartFormField
 import org.http4k.lens.MultipartFormFile
 import org.http4k.lens.Validator
 import org.http4k.lens.multipartForm
-import org.http4k.lens.string
 import org.http4k.server.SunHttp
 import org.http4k.server.asServer
 
@@ -46,7 +44,7 @@ fun main() {
     // taken care of automatically
     val multipartform = MultipartForm().with(
         nameField of Name("rita"),
-        imageFile of FormFile("image.txt", ContentType.OCTET_STREAM, "somebinarycontent".byteInputStream()))
+        imageFile of MultipartFormFile("image.txt", ContentType.OCTET_STREAM, "somebinarycontent".byteInputStream()))
     val validRequest = Request(Method.POST, "http://localhost:8000").with(strictFormBody of multipartform)
 
     println(ApacheClient()(validRequest))